--- apps/web/app/layout.tsx
@@ -1,30 +1,0 @@
-import type { Metadata } from 'next';
-import localFont from 'next/font/local';
-import './globals.css';
-
-const geistSans = localFont({
-  src: './fonts/GeistVF.woff',
-  variable: '--font-geist-sans',
-});
-const geistMono = localFont({
-  src: './fonts/GeistMonoVF.woff',
-  variable: '--font-geist-mono',
-});
-
-export const metadata: Metadata = {
-  description: 'Generated by create next app',
-  title: 'Create Next App',
-};
-const RootLayout = ({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) => (
-  <html lang="en">
-    <body className={`${geistSans.variable} ${geistMono.variable}`}>
-      {children}
-    </body>
-  </html>
-);
-
-export default RootLayout;
